fix: Moved JS-code to correct create.js
